+++ crates/compiler/src/evaluate/css_tree.rs
@@ -32,11 +32,11 @@ impl CssTree {
         tree
     }
 
-    pub fn get(&self, idx: CssTreeIdx) -> Ref<Option<CssStmt>> {
+    pub fn get(&self, idx: CssTreeIdx) -> Ref<'_, Option<CssStmt>> {
         self.stmts[idx.0].borrow()
     }
 
-    pub fn get_mut(&self, idx: CssTreeIdx) -> RefMut<Option<CssStmt>> {
+    pub fn get_mut(&self, idx: CssTreeIdx) -> RefMut<'_, Option<CssStmt>> {
         self.stmts[idx.0].borrow_mut()
     }