+++ crates/compiler/src/lib.rs
@@ -14,7 +14,7 @@ implementation.
 
 ## Use as library
 ```
-# use grass_compiler as grass;
+# use scss_compiler as grass;
 fn main() -> Result<(), Box<grass::Error>> {
     let css = grass::from_string(
         "a { b { color: &; } }".to_owned(),
@@ -32,7 +32,7 @@ grass input.scss
 ```
 */
 
-#![cfg_attr(doc_cfg, feature(doc_cfg))]
+#![cfg_attr(docsrs, feature(doc_cfg))]
 #![warn(clippy::all, clippy::cargo, clippy::dbg_macro)]
 #![deny(missing_debug_implementations)]
 #![allow(
@@ -225,7 +225,7 @@ fn from_string_with_file_name<P: AsRef<Path>>(
 /// n.b. `grass` does not currently support files or paths that are not valid UTF-8
 ///
 /// ```
-/// # use grass_compiler as grass;
+/// # use scss_compiler as grass;
 /// fn main() -> Result<(), Box<grass::Error>> {
 ///     let css = grass::from_path("input.scss", &grass::Options::default())?;
 ///     Ok(())
@@ -239,7 +239,7 @@ pub fn from_path<P: AsRef<Path>>(p: P, options: &Options) -> Result<String> {
 /// Compile CSS from a string
 ///
 /// ```
-/// # use grass_compiler as grass;
+/// # use scss_compiler as grass;
 /// fn main() -> Result<(), Box<grass::Error>> {
 ///     let css = grass::from_string("a { b { color: &; } }".to_string(), &grass::Options::default())?;
 ///     assert_eq!(css, "a b {\n  color: a b;\n}\n");