+++ crates/compiler/src/parse/sass.rs
@@ -81,7 +81,7 @@ impl<'a> StylesheetParser<'a> for SassParser<'a> {
         self.path
     }
 
-    fn options(&self) -> &Options {
+    fn options(&self) -> &'_ Options<'_> {
         self.options
     }